chemistryModel: Corrections to the Jacobian
The chemistry models have been corrected so that the temperature
derivative terms in the jacobian are correct and are consistent with the
derivatives method. This has resulted in a modest improvement in the
efficiency of the ODE integration.

A debug checking routine has been added to ODESystem which prints the
correspondence between the implemented jacobian and one constructed from
calling the derivatives method with finite differences. This will
facilitate convenient verification of the implementation in the future.

In addition a clean up of the derivatives and jacobian functions was
performed. Unused code has been removed and variables have been renamed
more meaningfully. Duplicate code has also been removed from the TDAC
chemistry model.