combustionModels/EDC: Added "outerCorrect" switch
This switch controls whether the model is corrected every outer
iteration, or just once per timestep.

This is the same as the corresponding switch in the laminar and PaSR
models, except that in the case of EDC it defaults to true. EDC folds
more information pertaining to transport into the chemistry integration,
so updating the chemistry on outer iteration is likely to have greater
effect than for the other models.