ISAT: Removed variable time step flag
This change ensures that ISAT operates correctly as a result of changes
to the chemistry integration step, by putting it permenantly into what
was previously a variable-time-step mode. The constant-time-step mode is
no longer available.

The constant-time-step mode was not compatible with run-time changes,
certain types of restart, or models in which chemistry is integrated
over a time-scale which differs from the physical time-step. The level
of optimsation that the constant-time-step mode previously provided was
also found to be negligible. It's removal therefore simplifies the code
and improves the correctness of operation whilst not removing any
meaningful functionality.