fvMeshSubset: Added new constraint patch type 'internal' into which exposed internal faces are added
This new constraint type is preferable to the 'empty' type used previously as it
support patch field values for post-processing and other purposes.

The internalFvPatchField operates as a 'zeroGradient' type so that the adjacent
cell values are displayed on the faces exposed by the sub-setting.

The internalFvsPatchField operates as a 'calculated' type so that the internal
face values are displayed on the faces exposed by the sub-setting.

The immediate benefit of this change can be seen when using 'subsetMesh' without
the '-noFields' option to create and write a sub-set of an 'fvMesh' with field
values, now the face values of the 'exposed' internal faces can be visualised.