MapFvSurfaceField: Generalised the mapping of surfaceField
to differentiate between flux field which require face-flipping and
non-extensive surface fields which do not.  Currently flux fields are
distinguished by being surfaceScalarField with dimensions of either volumetric
or mass flux.

This change corrects the handling of the surfaceVectorField Uf which was
previously mapped incorrectly on faces requiring the flipping of the flux
orientation.