tutorials/combustion/reactingFoam/laminar/counterFlowFlame2D(LTS): changed to Wilke transport mixing
Changed the laminar methane combustion cases to use the Wilke mixing rule for
the transport properties obtained from the Sutherland model but with coefficient
mixing for thermodynamic properties for efficient evaluation of reaction
equilibria.

This provides significantly more accurate results for laminar combustion,
producing a thinner flame and a 10K reduction in peak temperature.