waves: Optionally input period rather than length for Stokes-type waves
Stokes-type waves (i.e., Airy, Stokes2 and Stokes5) now support
specification in terms of wave period, as well as length. The following
specifications are therefore equivalent:

    Stokes2
    {
        length      300; // <-- Wave length specified
        amplitude   2.5;
        phase       0;
        angle       0;
        depth       50;
    }

    Stokes2
    {
        period      15.643; // <-- Wave period specified
        amplitude   2.5;
        phase       0;
        angle       0;
        depth       50;
    }

Note that the period is that of the wave in a reference frame with zero
mean flow. Specifying a doppler shifted period is not yet supported.

Exactly one of period or length must be specified. If both, or neither,
are provided then an error will be generated.

Corrections have also been made to the calculation of the celerity of
Stokes2 and Stokes5 waves.