topoSet: rotatedBoxToCell: Added box and rotation specification
The rotatedBoxToCell topoSet source can now be specified by providing
the geometry of a non-rotated box and how it should be rotated. For
example, in system/topoSetDict:

    actions
    (
        {
            name    c1;
            type    cellSet;
            action  new;
            source  rotatedBoxToCell;
            box     (0.4 0.4 -100) (0.6 0.6 100);
            centre  (0.4 0.4 0);
            n1      (1 0 0);
            n2      (1 1 0);
        }
    );

This selects cells in a box rotated from (1 0 0) to (1 1 0) (i.e., 45
degrees around the z-axis), with width and depth of 0.2, height of 200,
and with a bottom left corner at (0.4 0.4 -100).

The origin, i, j, k, specification is still available, and will be used
if the keyword "box" is not present. The equivalent input to the above
in this form is as follows:

    actions
    (
        {
            name    c1;
            type    cellSet;
            action  new;
            source  rotatedBoxToCell;
            origin  (0.4 0.4 -100);
            i       (0.141421 0.141421 0);
            j       (-0.141421 0.141421 0);
            k       (0 0 200);
        }
    );